added all blog categories and fixed table

+++ Makefile
@@ -1,5 +1,5 @@
 push:
-	rsync -avzr public/ lemon@joannalemon.com:/var/www/yulqen.org/ --exclude Makefile --exclude .git --delete-excluded --delete
+	hugo && rsync -avzr public/ lemon@joannalemon.com:/var/www/yulqen.org/ --exclude Makefile --exclude .git --delete-excluded --delete
 
 pull:
 	rsync -avzr lemon@joannalemon.com:/var/www/yulqen.org/ public/ --exclude Makefile --exclude .git --delete-excluded --delete
